Update PageContainer to become PageTemplate and include Mobile styling
