+++ src/pages/Home/index.tsx
@@ -1,14 +1,14 @@
 import { useIntl } from 'react-intl'
 
-import { PageContainer } from 'components/PageContainer'
+import { PageTemplate } from 'components/PageTemplate'
 
 import { MeowFact } from './MeowFact'
 
 export default () => {
   const { formatMessage } = useIntl()
   return (
-    <PageContainer heading={formatMessage({ id: 'home' })}>
+    <PageTemplate heading={formatMessage({ id: 'home' })}>
       <MeowFact />
-    </PageContainer>
+    </PageTemplate>
   )
 }
