+++ src/styles/Breakpoints.ts
@@ -1,0 +1,3 @@
+export const Breakpoints = {
+  Mobile: 'only screen and (max-width: 767px)',
+}
