+++ src/styles/index.ts
@@ -1,0 +1,5 @@
+import { Color as ColorImport } from 'styles/Colors'
+
+export type Color = ColorImport
+export { Breakpoints } from 'styles/Breakpoints'
+export { Colors } from 'styles/Colors'
